Use `git2` to read config values
When using an include directive in a .gitconfig file, manually parsing
the contents isn't enough to get all the configured variables. The
`git2` crate provides nice bindings around `libgit2`, enabling us to
interact with the config, which has resolved the include directives.

An additional benefit to using this crate, is the fact that it is no
longer necessary to pass the path to a `.gitconfig` from the command
line, as the `git2` crate provides an API to locate that file. This
improves usability significantly, as the user only has to run the exe,
without doing anything else.